Making it usable as a library.
